+++ src/components/reusable/sideDrawer/sideDrawer.scss
@@ -191,12 +191,6 @@ form {
         backdrop-filter: var(--kd-background-blur-medium);
       }
     }
-    &.no-backdrop {
-      &::backdrop {
-        background: transparent;
-        backdrop-filter: none;
-      }
-    }
   }
 }
 
@@ -210,6 +204,6 @@ form {
   }
 }
 
-.non-Modal {
+.no-backdrop {
   max-width: calc(100% - 2em - 6px);
 }
