--- src/components/reusable/widget/sample/gridstack.newWidget.scss
@@ -1,88 +1,0 @@
-@use '../../../../common/scss/utility/gridstack-shidoka.scss';
-
-:host {
-  display: block;
-}
-
-.test {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: var(--kd-color-background-container-soft);
-  height: 100%;
-  border-radius: 4px;
-}
-
-.new-items {
-  display: flex;
-}
-.hidden {
-  display: none;
-}
-
-.widget-content {
-  display: flex;
-  padding: 8px;
-  align-items: center;
-  gap: 16px;
-  align-self: stretch;
-  span > svg {
-    width: 80px;
-    height: 80px;
-  }
-}
-
-.dashboard-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 32px;
-  align-self: stretch;
-}
-.content-wrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 16px;
-  align-self: stretch;
-}
-.content-content {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  align-self: stretch;
-  margin-bottom: 32px;
-}
-.content-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  align-self: stretch;
-}
-.content-items {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-.content-item {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-}
-.widget-content {
-  display: flex;
-  padding: 8px;
-  align-items: center;
-  gap: 16px;
-  flex: 1 0 0;
-  align-self: stretch;
-  // min-height: 150px;
-  span > svg {
-    width: 80px;
-    height: 80px;
-  }
-}
-.star-filled > svg {
-  color: var(--kd-color-icon-brand);
-}
